Feature/finish requests refactor (#757)
* refactoring requests

* last cleanups

* last cleanups
