+++ r2r/main/api/routes/auth/base.py
@@ -3,9 +3,14 @@ from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from pydantic import BaseModel
 
 from r2r.base import Token, User, UserCreate
+from r2r.main.api.routes.auth.requests import (
+    PasswordChangeRequest,
+    PasswordResetConfirmRequest,
+    PasswordResetRequest,
+)
 
-from ...engine import R2REngine
-from .base_router import BaseRouter
+from ....engine import R2REngine
+from ..base_router import BaseRouter
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
@@ -16,19 +21,6 @@ class UserResponse(BaseModel):
 
 class TokenResponse(BaseModel):
     results: dict[str, Token]
-
-
-class PasswordChangeRequest(BaseModel):
-    current_password: str
-    new_password: str
-
-
-class PasswordResetRequest(BaseModel):
-    email: str
-
-
-class PasswordResetConfirmRequest(BaseModel):
-    new_password: str
 
 
 class UserProfileUpdate(BaseModel):
