+++ r2r/main/api/routes/auth/requests.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+
+class PasswordChangeRequest(BaseModel):
+    current_password: str
+    new_password: str
+
+
+class PasswordResetRequest(BaseModel):
+    email: str
+
+
+class PasswordResetConfirmRequest(BaseModel):
+    new_password: str
