+++ r2r/main/app.py
@@ -15,17 +15,24 @@ class R2RApp:
         uvicorn.run(self.app, host=host, port=port)
 
     def _setup_routes(self):
-        from .api.routes import auth, ingestion, management, retrieval
+        from .api.routes.auth import base as auth_base
+        from .api.routes.ingestion import base as ingestion_base
+        from .api.routes.management import base as management_base
+        from .api.routes.retrieval import base as retrieval_base
 
         self.app = FastAPI()
 
         # Create routers with the engine
-        ingestion_router = ingestion.IngestionRouter.build_router(self.engine)
-        management_router = management.ManagementRouter.build_router(
+        ingestion_router = ingestion_base.IngestionRouter.build_router(
             self.engine
         )
-        retrieval_router = retrieval.RetrievalRouter.build_router(self.engine)
-        auth_router = auth.AuthRouter.build_router(self.engine)
+        management_router = management_base.ManagementRouter.build_router(
+            self.engine
+        )
+        retrieval_router = retrieval_base.RetrievalRouter.build_router(
+            self.engine
+        )
+        auth_router = auth_base.AuthRouter.build_router(self.engine)
 
         # Include routers in the app
         self.app.include_router(ingestion_router, prefix="/v1")
